chore: move lib code
