+++ packages/eslint-config/src/lib/compat.ts
