+++ packages/eslint-config/src/lib/dir.ts
